Move GaussianPosterior and config dataclasses into gaussian_fullcov.py
GaussianPosterior and its config dataclasses (GaussianConfig, NetworkConfig,
OptimizerConfig, InferenceConfig) now live alongside GaussianFullCov in one
file. gaussian.py is reduced to the Gaussian factory function, which imports
what it needs from gaussian_fullcov.py. Public API is unchanged.
